+++ graphql/index.ts
@@ -44,6 +44,8 @@ import { MutateCourseResolver } from "./course/mutations";
 import { MutateConcreteNotificationsResolver } from "./concrete_notification/mutations";
 import { ExtendedFieldsConcreteNotificationResolver } from "./concrete_notification/fields";
 import { MutateSubcourseResolver } from "./subcourse/mutations";
+import { UserFieldsResolver } from "./user/fields";
+import { MutateUserResolver } from "./user/mutations";
 
 applyResolversEnhanceMap(authorizationEnhanceMap);
 applyResolversEnhanceMap(complexityEnhanceMap);
@@ -54,6 +56,8 @@ const schema = buildSchemaSync({
     resolvers: [
         /* User Authentication & Information */
         AuthenticationResolver,
+        UserFieldsResolver,
+        MutateUserResolver,
         FieldMeResolver,
         MutateMeResolver,
